LayoutTests/fast/frames/seamless/seamless-basic-expected.txt

PASS iframe.seamless is true
FAIL window.getComputedStyle(iframe).width should be 200px. Was -4px.
FAIL window.getComputedStyle(iframe).height should be 100px. Was 0px.
